requirements.txt
flask
flask-cors
flask-jwt-extended
pymongo
bcrypt
python-dotenv
scikit-learn
joblib
